[Gab] Correction du bug de rotation qui relançait l'application et changé de question + ajout d'un compteur de question sur l'interface graphique + ajout de ligne dans les questiosn réponses du fileloader

+++ ExamPreparation/app/src/main/java/com/example/gb/exampreparation/QuestionnerLauncherActivity.java
@@ -10,7 +10,7 @@ import android.widget.Button;
 public class QuestionnerLauncherActivity extends Activity {
 
     public static final String EXTRA = "ACTION";
-    public static final int CONTINUE = 1;
+    public static final byte CONTINUE = 1;
     public static final byte RESET = 2;
     public static final byte RELOAD = 3;
     public static final byte UNKNOWN = 4;
